+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0](https://www.npmjs.com/package/freeathome-local-api-client/v/1.3.0) - 18.06.2022
+
+### Added
+
+- [PG-189](https://pgerke.atlassian.net/browse/PG-189):
+  The `SystemAccessPoint` is now an `EventEmitter` emitting, for example, websocket events.
+
+### Fixed
+
+- [PG-193](https://pgerke.atlassian.net/browse/PG-193):
+  Rolled back to ts-node@10.8.0 to fix an issue with code coverage calculation.
+
 ## [1.2.0](https://www.npmjs.com/package/freeathome-local-api-client/v/1.2.0) - 16.06.2022
 
 ### Added